\name{teampackage}
\alias{teampackage-package}
\docType{package}
\title{package}
\description{package for team project}
\details{
10min_hourly_wind:Aggregates ten minute wind time series to hourly time series
Problem2:Medal comparison in Winter Olympics games
}
\author{Jonathan Ho, Jordan Mandel, Yue Wu <ywu67@jhu.edu>}
\keyword{package}
\examples{
problem1('bmw1Series',dnorm,c(-5,5),20,TRUE)
data(problem2)
}


